Change LZMA constants from c_int to u32, remove redundant casts

+++ liblzma-rs/src/lz/lz_encoder_mf.rs
@@ -1,5 +1,5 @@
 use crate::types::*;
-use core::ffi::{c_int, c_uint};
+use core::ffi::c_uint;
 extern "C" {
     static lzma_crc32_table: [[u32; 256]; 8];
 }
@@ -55,14 +55,14 @@ pub unsafe extern "C" fn lzma_mf_find(
     (*mf).read_ahead = (*mf).read_ahead.wrapping_add(1);
     return len_best;
 }
-pub const EMPTY_HASH_VALUE: c_int = 0;
+pub const EMPTY_HASH_VALUE: u32 = 0;
 pub const MUST_NORMALIZE_POS: c_uint = UINT32_MAX;
 unsafe extern "C" fn normalize(mf: *mut lzma_mf) {
-    let subvalue: u32 = (MUST_NORMALIZE_POS as u32).wrapping_sub((*mf).cyclic_size);
+    let subvalue: u32 = MUST_NORMALIZE_POS.wrapping_sub((*mf).cyclic_size);
     let mut i: u32 = 0;
     while i < (*mf).hash_count {
         if *(*mf).hash.offset(i as isize) <= subvalue {
-            *(*mf).hash.offset(i as isize) = EMPTY_HASH_VALUE as u32;
+            *(*mf).hash.offset(i as isize) = EMPTY_HASH_VALUE;
         } else {
             *(*mf).hash.offset(i as isize) =
                 (*(*mf).hash.offset(i as isize)).wrapping_sub(subvalue);
@@ -72,7 +72,7 @@ unsafe extern "C" fn normalize(mf: *mut lzma_mf) {
     let mut i_0: u32 = 0;
     while i_0 < (*mf).sons_count {
         if *(*mf).son.offset(i_0 as isize) <= subvalue {
-            *(*mf).son.offset(i_0 as isize) = EMPTY_HASH_VALUE as u32;
+            *(*mf).son.offset(i_0 as isize) = EMPTY_HASH_VALUE;
         } else {
             *(*mf).son.offset(i_0 as isize) =
                 (*(*mf).son.offset(i_0 as isize)).wrapping_sub(subvalue);
@@ -347,8 +347,8 @@ unsafe extern "C" fn bt_find_func(
         let fresh4 = depth;
         depth = depth.wrapping_sub(1);
         if fresh4 == 0 || delta >= cyclic_size {
-            *ptr0 = EMPTY_HASH_VALUE as u32;
-            *ptr1 = EMPTY_HASH_VALUE as u32;
+            *ptr0 = EMPTY_HASH_VALUE;
+            *ptr1 = EMPTY_HASH_VALUE;
             return matches;
         }
         let pair: *mut u32 = son.offset(
@@ -405,8 +405,8 @@ unsafe extern "C" fn bt_skip_func(
         let fresh5 = depth;
         depth = depth.wrapping_sub(1);
         if fresh5 == 0 || delta >= cyclic_size {
-            *ptr0 = EMPTY_HASH_VALUE as u32;
-            *ptr1 = EMPTY_HASH_VALUE as u32;
+            *ptr0 = EMPTY_HASH_VALUE;
+            *ptr1 = EMPTY_HASH_VALUE;
             return;
         }
         let pair: *mut u32 = son.offset(
